Extracted TvSearchModule & SearchIndexModule to their own classes
