+++ Ombi.UI/Ombi.UI.csproj
@@ -281,9 +281,11 @@
     <Compile Include="Modules\LandingPageModule.cs" />
     <Compile Include="Modules\LayoutModule.cs" />
     <Compile Include="Modules\SearchExtensionModule.cs" />
+    <Compile Include="Modules\Search\BaseSearchModule.cs" />
     <Compile Include="Modules\Search\MovieSearchModule.cs" />
     <Compile Include="Modules\Search\MusicSearchModule.cs" />
-    <Compile Include="Modules\Search\SearchModule.cs" />
+    <Compile Include="Modules\Search\SearchIndexModule.cs" />
+    <Compile Include="Modules\Search\TvSearchModule.cs" />
     <Compile Include="Modules\UserWizardModule.cs" />
     <Compile Include="NinjectModules\ApiModule.cs" />
     <Compile Include="NinjectModules\ConfigurationModule.cs" />
